Добавляет переключение табов
