+++ src/components/screen-my-list/screen-my-list.jsx
@@ -5,7 +5,7 @@ import MoviesList from "../movies-list/movies-list";
 import {validShape, validArrayOfShape} from "../../utils/props";
 
 const MyList = (props) => {
-  const {films, header, history} = props;
+  const {header, history, films} = props;
 
   return (
     <div className="user-page">
@@ -13,7 +13,7 @@ const MyList = (props) => {
 
       <section className="catalog">
         <h2 className="catalog__title visually-hidden">Catalog</h2>
-        <MoviesList films={films} />
+        <MoviesList moviesList={films} filmsRendered={films.length} />
       </section>
 
       <Footer />
