+++ src/index.js
@@ -7,12 +7,7 @@ import genres from "./mocks/genres";
 import {createStore} from "redux";
 import {Provider} from "react-redux";
 import {reducer} from "./store/reducer";
-
-const promoFilm = {
-  title: `The Grand Budapest Hotel`,
-  genre: `Drama`,
-  year: 2014
-};
+import {promoFilm} from "./utils/const";
 
 const store = createStore(
     reducer,
