+++ src/utils/const.js
@@ -1,0 +1,9 @@
+export const MORE_LIKE_NUM = 4;
+
+export const promoFilm = {
+  title: `The Grand Budapest Hotel`,
+  genre: `Drama`,
+  year: 2014
+};
+
+export const tabs = [`Overview`, `Details`, `Reviews`];
